refactor: mapper 와 generator 역할 클래스 패키지 분리
